blogs can now only be deleted by the user who added the blog

+++ part4/bloglist_backend/controllers/blogs.js
@@ -60,8 +60,24 @@ blogsRouter.put('/:id', async (request, response) => {
     response.json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
+blogsRouter.delete('/:id', tokenExtractor, async (request, response) => {
+    const decodedToken = jwt.verify(request.token, config.SECRET)
+    if (!decodedToken.id) {    
+        return response.status(401).json({ error: 'token invalid' })  
+    }
+    const user = await User.findById(decodedToken.id)
+    if (!user) {
+        return response.status(401).json({ error: 'user invalid' })  
+    }
+    
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+    }
+    else {
+        return response.status(401).json({ error: 'invalid user'})
+    }
+
     response.status(204).end()
 })
